Makefile update
- specific python version : python3.6 (current version of Tweepy not supported in 3.7 due to async/await reserved keywords)
- `make install` now avoids to overwrite config.yml

+++ Makefile
@@ -11,7 +11,7 @@ create-mastodon-cli:
 install:
 	test -d $(VENV) || virtualenv $(VENV) -p $(PYTHON_VERSION)
 	$(PIP) install -r $(REQUIREMENTS)
-	cp python-twootfeed/config.example.yml python-twootfeed/config.yml
+	test -e python-twootfeed/config.yml || cp python-twootfeed/config.example.yml python-twootfeed/config.yml
 
 lint:
 	$(PYTEST) --flake8 --isort -m "flake8 or isort" python-twootfeed
